+++ cookbooks/aws-parallelcluster-environment/attributes/environment.rb
@@ -5,6 +5,17 @@ default['cluster']['proxy'] = 'NONE'
 # For performance, set NFS threads to min(256, max(8, num_cores * 4))
 default['cluster']['nfs']['threads'] = [[node['cpu']['cores'].to_i * 4, 8].max, 256].min
 
+# NFS version configuration — these are attributes of the upstream `nfs` cookbook (cookbooks/nfs),
+# not PCluster-owned attributes. We override their defaults here to enforce NFSv4-only mode.
+# The upstream defaults are nil (defer to kernel), which leaves v2 and v3 enabled.
+default['nfs']['v4'] = 'yes'
+default['nfs']['v3'] = 'no'
+default['nfs']['v2'] = 'no'
+# PCluster-owned attribute: mask rpcbind.service and rpcbind.socket so systemd does not start
+# the auxiliary services (rpc-statd, nfs-mountd) that are only needed for NFSv2/v3.
+# Set to false if rpcbind is needed by other services on the node.
+default['cluster']['nfs']['mask-nfsv2-nfsv3-services'] = true
+
 # Kernel release version used to select Lustre version
 # This is a mechanism used to mock kernel release on docker system-tests, see kitchen.docker.yml:
 # when kernel_release is defined, it will be used, otherwise the release version will be taken from ohai.