+++ cookbooks/aws-parallelcluster-environment/resources/nfs/nfs_alinux2.rb
@@ -24,6 +24,7 @@ default_action :setup
 action :setup do
   action_install_nfs4
   action_disable_start_at_boot
+  action_mask_nfsv2_nfsv3_services
 end
 
 action_class do